[SYCL][NFC] Move reduction_forward header file
The file must be located in detail directory, which holds headers with
sycl::detail namespace definitions.
