+++ sycl/include/sycl/detail/reduction_forward.hpp
