+++ sycl/include/sycl/handler.hpp
@@ -17,6 +17,7 @@
 #include <sycl/detail/export.hpp>
 #include <sycl/detail/handler_proxy.hpp>
 #include <sycl/detail/os_util.hpp>
+#include <sycl/detail/reduction_forward.hpp>
 #include <sycl/event.hpp>
 #include <sycl/ext/intel/experimental/kernel_execution_properties.hpp>
 #include <sycl/ext/oneapi/device_global/device_global.hpp>
@@ -32,7 +33,6 @@
 #include <sycl/nd_item.hpp>
 #include <sycl/nd_range.hpp>
 #include <sycl/property_list.hpp>
-#include <sycl/reduction_forward.hpp>
 #include <sycl/sampler.hpp>
 #include <sycl/stl.hpp>
 #include <sycl/usm/usm_pointer_info.hpp>
